Move misc.js to js/
