+++ js/ffs/free.js
@@ -3,7 +3,7 @@ import $ from 'jquery';
 
 import '../promise-polyfill';
 import i18n from '../i18n';
-import {levenshteinDistance} from '../../libs/misc';
+import {levenshteinDistance} from '../misc';
 
 var freeFormQuery = {};
 var presets = {};
