+++ js/misc.js
